scraper/requirements.txt
selenium
pandas
numpy
boto3
python-dotenv
psycopg2-binary